#3778 separated RuleHistoryModal from ProjectRulesTab
